crm_service/requirements.txt
kafka-python>=2.0.0,<3
